pyproject.toml
[project]
name = "tcc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
    "streamlit>=1.45.0",
    "openai>=1.0.0",
]

[dependency-groups]
dev = [
    "docling>=2.57.0",
]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = "test_*.py"
python_functions = "test_*"
